fix : 로그인 로고 수정 및 view tsx 파일 생성
