+++ src/pages/SignIn/signIn.css
@@ -14,6 +14,7 @@
   display: table;
   border-radius: 50%;
   margin: auto;
+  overflow: hidden;
   text-align: center;
   background-color: white;
 }
@@ -28,9 +29,6 @@
   display: inline-block;
 }
 
-.logo > img {
-  vertical-align: middle;
-  position: relative;
-  width: 81%;
-  top: 10px;
+.logo-container {
+  top: 30px;
 }
